requirements.txt
async-timeout==4.0.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.2
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
pytz==2023.3.post1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
validators==0.22.0
Werkzeug==3.0.1
WTForms==3.1.0
